backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Config
pydantic-settings==2.6.1
python-dotenv==1.0.1

# Database
supabase==2.10.0
asyncpg==0.30.0

# PDF parsing
PyMuPDF==1.24.14

# HTTP client (async)
httpx==0.27.2

# LLM clients
groq==0.12.0
google-generativeai==0.8.3

# Utilities
python-multipart==0.0.12   # multipart file uploads
tenacity==9.0.0            # retry logic for API calls
numpy==1.26.4              # vector math

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
